Error Handling

+++ Application/AppTasks/Delete.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Application.Core;
+using Domain;
 using MediatR;
 using Persistence;
 
@@ -9,22 +11,28 @@ namespace Application.AppTasks
 {
     public class Delete
     {
-        public class Command : IRequest{
+        public class Command : IRequest<Result<Unit>>{
             public Guid Id { get; set; }
         }
 
-        public class Handle : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
         private readonly DataContext _context;
-            public Handle(DataContext context)
+            public Handler(DataContext context)
             {
             _context = context;
             }
 
-            async Task IRequestHandler<Command>.Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var appTask = await _context.AppTasks.FindAsync(request.Id);
+                if (appTask == null)
+                {
+                    return Result<Unit>.Failure("Item not found");
+                }
                 _context.AppTasks.Remove(appTask);
+                return Result<Unit>.Success(Unit.Value);
+                
             }
         }
     }
